+++ CHANGELOG.md
@@ -130,6 +130,7 @@ If you do, then the “Details” view contains the fg-graph plot and can tell y
 - When specifying a geometry file, white-space wasn’t stripped from the input field, creating unnecessary error conditions ([\#422](https://gitlab.desy.de/cfel-sc/amarcord-parent/amarcord-serial/-/issues/422))
 - The indexing-specific call to `list_events` used to fail because of some problems with temporary files. ([\#436](https://gitlab.desy.de/cfel-sc/amarcord-parent/amarcord-serial/-/issues/436), [\#441](https://gitlab.desy.de/cfel-sc/amarcord-parent/amarcord-serial/-/issues/441))
 - Editing the indexing command-line previously just added and changed options, and didn’t remove any ([\#447](https://gitlab.desy.de/cfel-sc/amarcord-parent/amarcord-serial/-/issues/447))
+- Runs API: you can now add new files to an existing run ([\#450](https://gitlab.desy.de/cfel-sc/amarcord-parent/amarcord-serial/-/issues/450))
 
 ## Development changes